requirements.txt
# Core automation
pywinauto>=0.6.8
pynput>=1.7.6
mss>=9.0.0
pytesseract>=0.3.10
opencv-python>=4.8.0
psutil>=5.9.0
pywin32>=306
Pillow>=10.0.0

# API server
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
websockets>=12.0

# AI providers
anthropic>=0.39.0
openai>=1.50.0
httpx>=0.27.0

# v2.0 - monitoring, tray, search, flows
pystray>=0.19.0
pyyaml>=6.0
duckduckgo_search>=7.0.0
playwright>=1.52.0

# v2.3 - recipe primitives (Phase 1), smart-home (Phase 7)
rapidfuzz>=3.9.0      # Fuzzy OCR label matching for gui_click_label
simpleeval>=0.9.13    # Safe expression eval for recipe skip_if
paho-mqtt>=2.1.0      # MQTT client for smart-home Phase 7
zeroconf>=0.132.0     # Hue / mDNS bridge discovery for Phase 7

# v2.3 - Phase 8 voice I/O (OPTIONAL server-side fallbacks)
# The primary voice path is the browser Web Speech API (zero-install,
# works in Edge/Chrome on Windows). These are only needed for Firefox,
# stripped WebViews, or users who want on-device transcription.
# Uncomment if you want server-side voice:
#
# faster-whisper>=1.0.0    # Local STT, CPU-friendly (~150 MB model)
# pyttsx3>=2.90            # Local TTS via Windows SAPI

